% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/first.R
\name{J}
\alias{J}
\title{Make a variable to be of Julia's awareness}
\usage{
J(x)
}
\description{
Make a variable to be of Julia's awareness, so it can be further used in
the definition of optimization problem.
}
\examples{
x <- Variable(4)
b <- J(c(1:4))
p <- minimize(sum((x - b) ^ 2))
}
