% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/first.R
\name{solve.problem}
\alias{solve.problem}
\title{Solve optimization problem}
\usage{
\method{solve}{problem}(p)
}
\arguments{
\item{p}{Optimization problem to be solved.}
}
\description{
\code{solve} solves optimization problem using Convex.jl.
}
\examples{
x <- Variable(4)
b <- J(c(1:4))
p <- minimize(sum((x - b) ^ 2))
solve(p)
}
