% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/first.R
\name{value}
\alias{value}
\alias{evaluate}
\title{Get values of expressions at optimizer}
\usage{
value(...)
}
\arguments{
\item{...}{Expressions needed to evaluate.}
}
\description{
\code{Value} returns the values of expressions at optimizer
(minimizer, maximizer and etc.).
}
\examples{
x <- Variable(4)
b <- J(c(1:4))
p <- minimize(sum((x - b) ^ 2))
cvx_optim(p)
value(x[1] + x[2], x[3] + x[4])
}
