% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/first.R
\name{cvx_optim}
\alias{cvx_optim}
\title{Solve optimization problem}
\usage{
cvx_optim(p)
}
\arguments{
\item{p}{optimization problem to be solved.}
}
\value{
status of optimized problem.
}
\description{
\code{cvx_optim} solves optimization problem using Convex.jl.
}
\examples{
x <- Variable(2)
b <- J(c(1:2))
p <- minimize(sum((x - b) ^ 2))
cvx_optim(p)
}
